# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Plant Diary API
  version: "0.1.0"
  description: 植物観察日記システムのAPI
servers:
  - url: http://localhost:8080
    description: ローカル開発サーバー
paths:
  /api/users:
    post:
      summary: ユーザーを作成する
      operationId: postApiUsers
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '503':
          description: Service Unavailable
        '500':
          description: Internal Server Error
  /api/photos:
    post:
      summary: 写真をアップロードする
      operationId: postApiPhotos
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadPhotoRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadPhotoResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '503':
          description: Service Unavailable
        '500':
          description: Internal Server Error
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    UserResponse:
      type: object
      required:
        - uuid
        - username
      properties:
        uuid:
          type: string
        username:
          type: string
    UploadPhotoRequest:
      type: object
      required:
        - photo
        - user_uuid
      properties:
        photo:
          type: string
          format: binary
        user_uuid:
          type: string
        captured_at:
          type: string
          format: date-time
    UploadPhotoResponse:
      type: object
      required:
        - job_id
      properties:
        job_id:
          type: string